docs : page 연결 구조 변경 및 components 수정
