+++ src/App.js
@@ -3,23 +3,17 @@ import Intro from "./pages/Intro";
 import Index from "./pages/Index";
 import { Wrap } from "./styles/basic";
 import { Route, Routes } from "react-router";
-// import RecipeWrite from "./pages/RecipeWrite";
 import RecipeMore from "./pages/RecipeMore";
-import MealBig from "./pages/home/MealBig";
-import MealSmall from "./pages/home/MealSmall";
 import NotFound from "./pages/NotFound";
 
 const App = () => {
   return (
     <Wrap width={320} minh={100}>
       <Routes>
-        <Route path="/" element={<Intro />}></Route>
+        <Route path="/" element={<Intro />} />
         <Route path="/meal" element={<Index />}></Route>
-        <Route path="/meal/big" element={<MealBig />}></Route>
-        <Route path="/meal/small" element={<MealSmall />}></Route>
-        {/* <Route path="/meal/write" element={<RecipeWrite />}></Route> */}
-        <Route path="/meal/more" element={<RecipeMore />}></Route>
-        <Route path="*" element={<NotFound />}></Route>
+        <Route path="/meal/more" element={<RecipeMore />} />
+        <Route path="*" element={<NotFound />} />
       </Routes>
     </Wrap>
   );
