+++ src/styles/meal/noticeboard.js
@@ -1,15 +1,9 @@
 import styled from "@emotion/styled";
 
 export const NoticeBoardDiv = styled.div`
-  display: flex;
-  flex-wrap: wrap;
-  justify-content: space-around;
-`;
-
-export const NoticeBoardStyle = styled.div`
   width: 128px;
   margin: 15px 0 0;
-  background: #ffebd0;
+  background: var(--brand-color2);
   padding: 5px;
   box-shadow: 2px 2px 5px 0px rgba(0, 0, 0, 0.2);
 `;
@@ -34,12 +28,11 @@ export const NoticeBoardTags = styled.div`
   position: relative;
   display: flex;
   flex-wrap: wrap;
-  gap: 5px;
-  margin-bottom: 10px;
+  gap: 3px;
+  margin-bottom: 3px;
   span {
-    border: none;
     background: #fff;
-    border-radius: 5px;
+    border-radius: 10px;
     padding: 3px;
     font-size: 0.7rem;
   }
